Relocates fixtures from conftest.py: Moves fixtures that were only referenced by a single test module out of conftest.py and into their respective test modules
