+++ app/tests/integration_tests/citibuy/test_client.py
@@ -1,6 +1,14 @@
 import pytest
 import sqlalchemy
 
+from dgs_fiscal.systems import CitiBuy
+
+
+@pytest.fixture(scope="session", name="test_citibuy")
+def fixture_citibuy():
+    """Creates an instance of the CitiBuy class for integration testing"""
+    return CitiBuy()
+
 
 class TestCitiBuy:
     """Tests the CitiBuy class"""
